// @types/generated/contentful.d.ts
export interface IAbTestFields {
  /** List Name */
  name: string;

  /** A/B Test */
  unfrmAbTest: Record<string, any>;

  /** Variants */
  variants: IAbTestVariant[];

  /** Personalization Criteria */
  unfrmOptPersonalizationCriteria?: Record<string, any> | undefined;
}

export interface IAbTest extends Entry<IAbTestFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ABTest";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IAbTestVariantFields {
  /** Identifier */
  identifier?: string | undefined;

  /** Distribution */
  distribution?: number | undefined;

  /** Component */
  component?: Entry<{ [fieldId: string]: unknown }> | undefined;
}

export interface IAbTestVariant extends Entry<IAbTestVariantFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "abTestVariant";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}